# docs/architecture/api_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "A股/港股/美股自选股智能分析系统 API\n\n## 功能模块\n- 股票分析：触发 AI 智能分析\n- 历史记录：查询历史分析报告\n- 股票数据：获取行情数据\n\n## 认证方式\n当前版本暂无认证要求",
    "version": "1.0.0",
    "contact": {
      "name": "Daily Stock Analysis Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
...
    "/api/v1/history": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "获取历史分析列表",
        "description": "分页获取历史分析记录摘要，支持按股票代码和日期范围筛选",
        "operationId": "getHistoryList",
        "parameters": [
          {
            "name": "stock_code",
            "in": "query",
            "description": "股票代码筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "description": "开始日期 (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "description": "结束日期 (YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码（从 1 开始）",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "历史记录列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{query_id}": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "获取历史报告详情",
        "description": "根据 query_id 获取完整的历史分析报告",
        "operationId": "getHistoryDetail",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "description": "分析记录唯一标识",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "报告详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "404": {
            "description": "报告不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{query_id}/news": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "获取历史报告关联新闻",
        "description": "根据 query_id 获取关联的新闻情报列表（为空也返回 200）",
        "operationId": "getHistoryNews",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "description": "分析记录唯一标识",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返回数量限制",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "新闻情报列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsIntelResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HistoryListResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "总记录数"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryItem"
            }
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "items"
        ]
      },
      "NewsIntelItem": {
        "type": "object",
        "description": "新闻情报条目",
        "properties": {
          "title": {
            "type": "string",
            "description": "新闻标题"
          },
          "snippet": {
            "type": "string",
            "description": "新闻摘要（最多50字）"
          },
          "url": {
            "type": "string",
            "description": "新闻链接"
          }
        },
        "required": [
          "title",
          "url"
        ]
      },
      "NewsIntelResponse": {
        "type": "object",
        "description": "新闻情报响应",
        "properties": {
          "total": {
            "type": "integer",
            "description": "新闻条数"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsIntelItem"
            },
            "description": "新闻列表"
          }
        },
        "required": [
          "total",
          "items"
        ]
      },
      "HistoryItem": {
        "type": "object",
        "description": "历史记录摘要（列表展示用）",
        "properties": {
          "query_id": {
            "type": "string"
          },
          "stock_code": {
            "type": "string"
          },
          "stock_name": {
            "type": "string"
          },
          "report_type": {
            "type": "string"
          },
          "sentiment_score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "operation_advice": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "query_id",
          "stock_code",
          "created_at"
        ]
      },
      "AnalysisReport": {
        "type": "object",
        "description": "完整分析报告",
        "properties": {
          "meta": {
            "type": "object",
            "description": "元信息",
            "properties": {
              "query_id": {
                "type": "string"
              },
              "stock_code": {
                "type": "string"
              },
              "stock_name": {
                "type": "string"
              },
              "report_type": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "model_used": {
                "type": "string",
                "description": "分析使用的 LLM 模型（完整名，如 gemini/gemini-2.0-flash）"
              }
            }
          },
          "summary": {
            "type": "object",
            "description": "概览区（首屏展示）",
            "properties": {
              "analysis_summary": {
                "type": "string",
                "description": "关键结论"
              },
              "operation_advice": {
                "type": "string",
                "description": "操作建议"
              },
              "trend_prediction": {
                "type": "string",
                "description": "趋势预测"
              },
              "sentiment_score": {
                "type": "integer",
                "description": "情绪评分 (0-100)"
              },
              "sentiment_label": {
                "type": "string",
                "description": "情绪标签",
                "enum": [
                  "极度悲观",
                  "悲观",
                  "中性",
                  "乐观",
                  "极度乐观"
                ]
              }
            }
          },
          "strategy": {
            "type": "object",
            "description": "策略点位区",
            "properties": {
              "ideal_buy": {
                "type": "string",
                "description": "理想买入价"
              },
              "secondary_buy": {
                "type": "string",
                "description": "第二买入价"
              },
              "stop_loss": {
                "type": "string",
                "description": "止损价"
              },
              "take_profit": {
                "type": "string",
                "description": "止盈价"
              }
            }
          },
          "details": {
            "type": "object",
            "description": "详情区（可折叠）",
            "properties": {
              "news_content": {
                "type": "string",
                "description": "新闻摘要"
              },
              "raw_result": {
                "type": "object",
                "description": "原始分析结果（JSON）"
              },
              "context_snapshot": {
                "type": "object",
                "description": "分析时上下文快照（JSON）"
              },
              "stock_profile": {
                "type": "object",
                "description": "Deep Research 生成的股票基本情况"
              }
            }
          }
        },
        "required": [
          "meta",
          "summary"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误类型"
          },
          "message": {
            "type": "string",
            "description": "错误详情"
          },
          "detail": {
            "type": "object",
            "description": "附加错误信息"
          }
        },
        "required": [
          "error",
          "message"
        ]
      }
    }
  }
}